% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_df.R
\name{as_df}
\alias{as_df}
\title{Convert to a data.frame}
\usage{
as_df(x)
}
\arguments{
\item{x}{an object of class \code{bhl_meta} or \code{bhl_meta_single}.
\code{\link{sp_bhl_meta}} outputs an object of class \code{bhl_meta}, composed
of objects of class \code{bhl_meta_single}.}
}
\value{
If \code{bhl_meta} a single data.frame. If \code{bhl_meta_single},
a list of data.frame's of length equal to that of the input list.
}
\description{
Convert to a data.frame
}
\examples{
\dontrun{
geom <- 'POLYGON((-124.07 41.48,-119.99 41.48,-119.99 35.57,-124.07 35.57,-124.07 41.48))'
res <- sp_occ_gbif(geometry = geom)
res \%>\% sp_list()
x <- res \%>\% sp_list() \%>\% .[1:2] \%>\% sp_bhl_meta()

##
as_df(x$`allium amplectens`)
as_df(x$`allium falcifolium`)
as_df(x)
}
}

