% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spplit-package.R
\docType{data}
\name{gbif_datasets}
\alias{gbif_datasets}
\title{Named list of length seven}
\description{
Each element named with the California Academy of Sciences collection
name
}
\details{
Each element has one character string that represents the
collection codes for CAS in the GBIF database. These codes can be
used to search GBIF.
}
\keyword{data}
