% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spplit-package.R
\docType{data}
\name{idigbio_recordsets}
\alias{idigbio_recordsets}
\title{Named list of length seven}
\description{
Each element named with the California Academy of Sciences collection
name
}
\details{
Each element has one to many character strings that represent the
collection codes for CAS in the iDigBio database. These codes can be
used to search iDigBio.
}
\keyword{data}

