% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_authors.R
\name{sp_authors}
\alias{sp_authors}
\title{Get an author list (unique set of authors)}
\usage{
sp_authors(x, which = "recorded")
}
\arguments{
\item{x}{Output from a call to \code{\link{sp_occ_idigbio}} or \code{\link{sp_occ_gbif}}}
}
\value{
gives a character vector of unique taxon list
}
\description{
Get an author list (unique set of authors)
}
\details{
only accepts objects of class \code{occdatind} right now
}
\examples{
\dontrun{
# idigbio
geom <- 'POLYGON((-124.07 41.48,-119.99 41.48,-119.99 35.57,-124.07 35.57,-124.07 41.48))'
res <- sp_occ_idigbio(geometry = geom)
res \%>\% sp_authors()

# gbif
geom <- 'POLYGON((-124.07 41.48,-119.99 41.48,-119.99 35.57,-124.07 35.57,-124.07 41.48))'
res <- sp_occ_gbif(geometry = geom)
res \%>\% sp_authors()
}
}

