% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_bhl_ocr.R
\name{sp_bhl_ocr}
\alias{sp_bhl_ocr}
\title{Get BHL OCR text}
\usage{
sp_bhl_ocr(x)
}
\arguments{
\item{x}{input}
}
\description{
Get BHL OCR text
}
\examples{
\dontrun{
geom <- 'POLYGON((-124.07 41.48,-119.99 41.48,-119.99 35.57,-124.07 35.57,-124.07 41.48))'
res <- sp_occ(geometry = geom)
res \%>\% sp_list()
x <- res \%>\% sp_list() \%>\% sp_bhl_meta()
x[1:3] \%>\% sp_bhl_ocr
}
}

