% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_bhl_save.R
\name{sp_bhl_save}
\alias{sp_bhl_save}
\title{Save OCR corpus}
\usage{
sp_bhl_save(x)
}
\arguments{
\item{x}{input}
}
\description{
Save OCR corpus
}
\examples{
\dontrun{
geom <- 'POLYGON((-124.07 41.48,-119.99 41.48,-119.99 35.57,-124.07 35.57,-124.07 41.48))'
res <- sp_occ(geometry = geom, limit = 5)
x <- res \%>\% sp_list() \%>\% sp_bhl_meta() \%>\% .[1:10] \%>\% sp_bhl_ocr()
x <- unlist(x, FALSE)
x \%>\% sp_bhl_save()
}
}

