% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_lit_meta.R
\name{sp_lit_meta}
\alias{sp_lit_meta}
\title{Get literature metadata based on a species list}
\usage{
sp_lit_meta(x, from = "pubmed", limit = 25, progress = TRUE)
}
\arguments{
\item{x}{input}

\item{from}{(character) source to search for literature. one of more of
pubmed, bhl. default: pubmed}

\item{limit}{(integer/numeric) number of literature search results to
return}

\item{progress}{(logical) print a progress bar. default: \code{TRUE}}
}
\value{
a list
}
\description{
Get literature metadata based on a species list
}
\details{
See \link{spplit_auth} for authentication
}
\examples{
\dontrun{
# occurrence data
library(spocc)
geom <- 'POLYGON((-124.07 41.48,-119.99 41.48,-119.99 35.57,-124.07 35.57,-124.07 41.48))'
res <- occ(geometry = geom, from = "gbif", limit = 15)

taxa <- c('Pinus contorta', 'Accipiter striatus')
res <- occ(query=taxa, from = c("gbif", "bison"), limit = 15)
res
res$gbif
res$bison

# get literature metadata
x <- sp_lit_meta(x = res, from = "pubmed")
x
x[[1]]
x[[3]]
}
}
