% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_lit_names.R
\name{sp_lit_names}
\alias{sp_lit_names}
\title{Extract scientific names from articles}
\usage{
sp_lit_names(x, progress = TRUE, ...)
}
\arguments{
\item{x}{An object of class \code{sp_lit_text} or \code{sp_lit_text_one}}

\item{progress}{(logical) print a progress bar. default: \code{TRUE}}

\item{...}{arguments passed on to \code{namext::name_extract()}}
}
\value{
a data.frame
}
\description{
Depends on non-CRAN package namext
}
\details{
make sure to install gnfinder first
\url{https://github.com/gnames/gnfinder}
}
\examples{
\dontrun{
geom <- 'POLYGON((-124.07 41.48,-119.99 41.48,-119.99 35.57,-124.07 35.57,-124.07 41.48))'
res <- sp_occ(geometry = geom, from = "gbif", limit = 15)
spp <- sp_spp(res)
z <- sp_lit_text(spp)
z
z[1]
sp_lit_names(z[1])
sp_lit_names(z)
}
}
