% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_lit_parse_names.R
\name{sp_lit_parse_names}
\alias{sp_lit_parse_names}
\title{Parse scientific names}
\usage{
sp_lit_parse_names(x, progress = TRUE, ...)
}
\arguments{
\item{x}{An object of class \code{sp_lit_text} or \code{sp_lit_text_one}}

\item{progress}{(logical) print a progress bar. default: \code{TRUE}}

\item{...}{arguments passed on to \code{rgnparser::gn_parse_tidy()}}
}
\value{
a data.frame
}
\description{
Depends on non-CRAN package rgnparser
}
\details{
make sure to install gnparser first
\url{https://gitlab.com/gogna/gnparser}
}
