% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_lit_text.R
\name{sp_lit_text}
\alias{sp_lit_text}
\title{Fetch full text of articles}
\usage{
sp_lit_text(x, from = "pubmed", progress = TRUE)
}
\arguments{
\item{x}{An object of class \code{lit_meta} or \code{lit_meta_one}}

\item{from}{(character) one or more of: bhl, pubmed (default: pubmed)}

\item{progress}{(logical) print a progress bar. default: \code{TRUE}}
}
\value{
An object of class \code{lit_text}, or a list of such objects
}
\description{
Fetch full text of articles
}
\details{
See \link{spplit_auth} for authentication
}
\examples{
\dontrun{
library(spocc)
taxa <- c('Pinus contorta', 'Accipiter striatus')
res <- occ(query=taxa, from = c("gbif", "bison"), limit = 15)
w <- sp_lit_meta(x = res, from = c("pubmed", "bhl"))
w
w[[1]]$bhl
# x = w[[1]]; from = c("pubmed", "bhl")
out <- sp_lit_text(x = w[[1]], from = c("pubmed", "bhl"))
out
}
}
