% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sl_xy.R
\name{sp_occ}
\alias{sp_occ}
\title{Explore what data is available}
\usage{
sp_occ(query = NULL, geometry = NULL, limit = 10, ...)
}
\arguments{
\item{query}{Scientific names, optional}

\item{geometry}{WKT or bounding box, optional}

\item{limit}{Number results to return}

\item{...}{Curl options passed to \code{\link[httr]{GET}}}
}
\description{
Explore what data is available
}
\examples{
\dontrun{
geom <- 'POLYGON((-124.07 41.48,-119.99 41.48,-119.99 35.57,-124.07 35.57,-124.07 41.48))'
res <- sp_occ(geometry = geom)
res \%>\% sp_list()
x <- res \%>\% sp_list() \%>\% sp_bhl_meta()
ocred <- x[1:10] \%>\% sp_bhl_ocr
ocred \%>\% sp_bhl_save()
#res \%>\% sp_list() \%>\% sp_bhl_meta() \%>\% sp_bhl_ocr \%>\% sp_bhl_save()
}
}

