% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_occ_idigbio.R
\name{sp_occ_idigbio}
\alias{sp_occ_idigbio}
\title{Search for occurrences in iDigBio}
\usage{
sp_occ_idigbio(
  query = NULL,
  geometry = NULL,
  limit = 10,
  cas_coll = "botany",
  args = NULL,
  ...
)
}
\arguments{
\item{query}{(character) Scientific names, optional}

\item{geometry}{WKT or bounding box, optional}

\item{limit}{(integer) Number results to return}

\item{cas_coll}{(character) CAS collection name OR a collection code. See Details.}

\item{args}{additional args to iDigBio, see
\url{https://github.com/iDigBio/idigbio-search-api/wiki/Index-Fields} and
\url{https://github.com/iDigBio/idigbio-search-api/wiki/Query-Format}}

\item{...}{Further options passed to \code{\link[spocc:occ]{spocc::occ()}}}
}
\description{
Search for occurrences in iDigBio
}
\details{
This has hard-coded internal settings to get data from the
Cal Academy of Sciences collections within iDigBio
}
\section{cas_coll options}{

\itemize{
\item entomology
\item herpetology
\item ichthyology
\item invertebrate zoology & geology
\item ornithology
\item mammalogy
\item botany
}

OR a recordset, see \link{idigbio_recordsets}
}

\examples{
\dontrun{
geom <- 'POLYGON((-124.07 41.48,-119.99 41.48,-119.99 35.57,-124.07 35.57,-124.07 41.48))'
res <- sp_occ_idigbio(geometry = geom)
res \%>\% sp_list()
x <- res \%>\% sp_list() \%>\% sp_bhl_meta()
ocred <- x[1:10] \%>\% sp_bhl_ocr
ocred \%>\% sp_bhl_save()
#res \%>\% sp_list() \%>\% sp_bhl_meta() \%>\% sp_bhl_ocr \%>\% sp_bhl_save()

# geometry and class arachnida
sp_occ_idigbio(geometry = geom, args = c(order = 'Asterales'))
sp_occ_idigbio(geometry = geom, args = c(order = 'squamata'), cas_coll = "herpetology")

# just class arachnida - FIXME, no results
library("httr")
# sp_occ_idigbio(args = c(class='arachnida'), cas_coll = "entomology", callopts=verbose())

# specify CAS collection (default: botany)
## single
sp_occ_idigbio(geometry = geom, cas_coll = "entomology")
sp_occ_idigbio(geometry = geom, cas_coll = "botany")
sp_occ_idigbio(geometry = geom, cas_coll = "herpetology")
## all collections
sp_occ_idigbio(geometry = geom, cas_coll = "all")
## multiple collections
sp_occ_idigbio(geometry = geom, cas_coll = c("entomology", "herpetology"))
}
}
