% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_plos_fetch.R
\name{sp_plos_fetch}
\alias{sp_plos_fetch}
\title{Get PLOS full text}
\usage{
sp_plos_fetch(x, progess = TRUE, ...)
}
\arguments{
\item{x}{An object of class \code{plos_meta} or \code{plos_meta_single}}

\item{...}{curl options passed on to \code{\link[httr]{GET}}}

\item{progress}{(logical) print a progress bar. default: \code{TRUE}}
}
\description{
Get PLOS full text
}
\details{
Uses \code{\link[rplos]{plos_fulltext}} to fetch full text XML
}
\examples{
\dontrun{
geom <- 'POLYGON((-124.07 41.48,-119.99 41.48,-119.99 35.57,-124.07 35.57,-124.07 41.48))'
res <- sp_occ_gbif(geometry = geom, limit = 50)
z <- res \%>\% sp_list() \%>\% sp_plos_meta()
sp_plos_fetch(z[[1]])
sp_plos_fetch(z[1:3])

txt <- sp_plos_fetch(z[[3]])
txt$`angelica californica`
txt$`angelica californica`$`10.1371/journal.pone.0092265`
}
}

