% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_plos_meta.R
\name{sp_plos_meta}
\alias{sp_plos_meta}
\title{Get PLOS metadata based on a species list}
\usage{
sp_plos_meta(x, progress = TRUE, ...)
}
\arguments{
\item{x}{(character/list) A vector or list of character strings to search on}

\item{progress}{(logical) print a progress bar. default: \code{TRUE}}

\item{...}{curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Get PLOS metadata based on a species list
}
\section{Defaults}{

A set of defaults are used, based on what we think are very reasonable
assumptions about what most people want:
\itemize{
 \item Full research articles - we assume most people want full articles, not subsets,
 comments, etc. - toggle this by xxx
 \item Fields: we fetch the fields \code{id, title, authors}, but you can specify
 which you'd like back by xxx
}
}

\section{Internals}{

Internally, this function loops over your inputs vector list passed in to the
parameter \code{x}. With each loop, we use \code{\link[rplos]{searchplos}}
with above defaults.
}
\examples{
\dontrun{
geom <- 'POLYGON((-124.07 41.48,-119.99 41.48,-119.99 35.57,-124.07 35.57,-124.07 41.48))'
res <- sp_occ_gbif(geometry = geom, limit = 300)

# get species list first, then pass to sp_plos_meta
res \%>\% sp_list() \%>\% sp_plos_meta()

# or, pass directly to sp_plos_meta, and species list is extracted automatically
res \%>\% sp_plos_meta()

# combine all into a data.frame
as_df(z$`allium amplectens`)
df <- as_df(z)
names(df)
df$title
}
}

