% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_pubmed_fetch.R
\name{sp_pubmed_fetch}
\alias{sp_pubmed_fetch}
\title{Get Pubmed articles}
\usage{
sp_pubmed_fetch(x, key = NULL, progress = TRUE)
}
\arguments{
\item{x}{An object of class \code{pubmed_meta} or \code{pubmed_meta_single}}

\item{key}{api key, required. Go to
\url{http://www.biodiversitylibrary.org/getapikey.aspx}
to get a key. you can pass in as a parameter here, or leave \code{NULL} and
store as an R option or env variable. See \strong{BHL Authentication}
section in \link{spplit}}

\item{progress}{(logical) print a progress bar. default: \code{TRUE}}
}
\value{
An object of class \code{pubmed_fetch}, or a list of such objects
}
\description{
Get Pubmed articles
}
\examples{
\dontrun{
geom <- 'POLYGON((-124.07 41.48,-119.99 41.48,-119.99 35.57,-124.07 35.57,-124.07 41.48))'
res <- sp_occ_gbif(geometry = geom, limit = 15)
spp <- sp_spp(res)
z <- sp_pubmed_meta(spp)
z
z[1]
x <- sp_pubmed_fetch(z[1])
sp_pubmed_fetch(z)
}
}
