% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_pubmed_meta.R
\name{sp_pubmed_meta}
\alias{sp_pubmed_meta}
\title{Get Pubmed metadata based on a species list}
\usage{
sp_pubmed_meta(x, limit = 25, key = NULL, progress = TRUE)
}
\arguments{
\item{x}{input}

\item{limit}{(integer/numeric) number of literature search results to
return}

\item{key}{api key, required. Go to
\url{http://www.biodiversitylibrary.org/getapikey.aspx}
to get a key. you can pass in as a parameter here, or leave \code{NULL} and
store as an R option (as \code{bhl_key}) or environment variable (as
\code{BHL_KEY}). See \strong{BHL Authentication} section in \link{spplit} for more}

\item{progress}{(logical) print a progress bar. default: \code{TRUE}}
}
\value{
a list
}
\description{
Get Pubmed metadata based on a species list
}
\examples{
\dontrun{
geom <- 'POLYGON((-124.07 41.48,-119.99 41.48,-119.99 35.57,-124.07 35.57,-124.07 41.48))'
res <- sp_occ_gbif(geometry = geom, limit = 15)
spp <- sp_spp(res)
x <- sp_pubmed_meta(spp)
x
x[[1]]
x[[3]]
}
}
