% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sl_x.R
\name{sp_search}
\alias{sp_search}
\title{Explore what data is available}
\usage{
sp_search(geometry = NULL, limit = 10)
}
\arguments{
\item{variable}{Query for a variable}

\item{source}{Query for a data source}
}
\description{
Explore what data is available
}
\examples{
\dontrun{
geom <- 'POLYGON((-124.07 41.48,-119.99 41.48,-119.99 35.57,-124.07 35.57,-124.07 41.48))'
res <- sp_search(geometry = geom)
}
}

