% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spplit-package.R
\name{spplit_auth}
\alias{spplit_auth}
\title{Authentication}
\description{
Authentication
}
\section{BHL Authentication}{

For access to Biodiveristy Heritage Library data, you'll need an API key from them.
To get one fill out the brief form at \url{http://www.biodiversitylibrary.org/getapikey.aspx} -
they'll ask for your name and email address.

To use the key, do one of:
\itemize{
\item pass it in the \code{key} parameter in \code{\link[=sp_bhl_meta]{sp_bhl_meta()}} and \code{\link[=sp_bhl_ocr]{sp_bhl_ocr()}}
\item store as an environment variable (as \code{BHL_KEY}) either in your \code{.Renviron}
file, or wherever you store your environment variables (e.g., \code{.bashrc}, or
\code{.bash_profile}, or \code{.zshrc})
\item store as an R option (as \code{bhl_key}) in your \code{.Rprofile} file
}
}

