% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewer.R
\name{viewer}
\alias{viewer}
\title{Browse highlighted fragments in your default browser}
\usage{
viewer(input = NULL, output = NULL, browse = TRUE)
}
\arguments{
\item{input}{Input}

\item{output}{Path and file name for output file. If NULL, a temp file is used.}

\item{browse}{Browse file in your default browse immediately after file creation.
If FALSE, the file is written, but not opened.}
}
\description{
Browse highlighted fragments in your default browser
}
\examples{
\dontrun{
viewer(out)
}
}

