% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_bhl_meta.R
\name{sp_bhl_meta}
\alias{sp_bhl_meta}
\title{Get BHL metadata based on a species list}
\usage{
sp_bhl_meta(x)
}
\arguments{
\item{x}{input}
}
\description{
Get BHL metadata based on a species list
}
\examples{
\dontrun{
geom <- 'POLYGON((-124.07 41.48,-119.99 41.48,-119.99 35.57,-124.07 35.57,-124.07 41.48))'
res <- sp_occ(geometry = geom)
res \%>\% sp_list()
x <- res \%>\% sp_list() \%>\% sp_bhl_meta()
}
}

