% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_bhl_meta.R
\name{sp_bhl_meta}
\alias{sp_bhl_meta}
\title{Get BHL metadata based on a species list}
\usage{
sp_bhl_meta(x, key = NULL, progress = TRUE)
}
\arguments{
\item{x}{input}

\item{key}{api key, required. Go to
\url{http://www.biodiversitylibrary.org/getapikey.aspx}
to get a key. you can pass in as a parameter here, or leave \code{NULL} and
store as an R option (as \code{bhl_key}) or environment variable (as
\code{BHL_KEY}). See \strong{BHL Authentication} section in \link{spplit} for more}

\item{progress}{(logical) print a progress bar. default: \code{TRUE}}
}
\value{
a list
}
\description{
Get BHL metadata based on a species list
}
\examples{
\dontrun{
geom <- 'POLYGON((-124.07 41.48,-119.99 41.48,-119.99 35.57,-124.07 35.57,-124.07 41.48))'
res <- sp_occ_gbif(geometry = geom)
res \%>\% sp_spp()
x <- res \%>\% sp_spp() \%>\% .[1:2] \%>\% sp_bhl_meta()

# combine all into a data.frame
# FIXME: doesn't work
# as_df(x$`bolboschoenus maritimus`)
# as_df(x)

# with collector names/authors
geom <- 'POLYGON((-124.07 41.48,-119.99 41.48,-119.99 35.57,-124.07 35.57,-124.07 41.48))'
res <- sp_occ_gbif(geometry = geom, limit = 50)
authors <- res \%>\% sp_authors()
outx <- structure(authors[1:15], class="spauthors") \%>\% sp_bhl_meta()
outx
outx$`cummings, alice`
outx$`wagner, d. h.`
}
}
