% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_bhl_ocr.R
\name{sp_bhl_ocr}
\alias{sp_bhl_ocr}
\title{Get BHL OCR text}
\usage{
sp_bhl_ocr(x, key = NULL, progress = TRUE)
}
\arguments{
\item{x}{An object of class \code{bhl_meta} or \code{bhl_meta_single}}

\item{key}{api key, required. Go to \url{http://www.biodiversitylibrary.org/getapikey.aspx}
to get a key. you can pass in as a parameter here, or leave \code{NULL} and store as
an R option or env variable. See \strong{BHL Authentication} section in \link{spplit}}

\item{progress}{(logical) print a progress bar. default: \code{TRUE}}
}
\value{
An object of class \code{bhl_ocr}, or a list of such objects
}
\description{
Get BHL OCR text
}
\examples{
\dontrun{
geom <- 'POLYGON((-124.07 41.48,-119.99 41.48,-119.99 35.57,-124.07 35.57,-124.07 41.48))'
res <- sp_occ_gbif(geometry = geom)
(x <- res \%>\% sp_list() \%>\% .[1:2] \%>\% sp_bhl_meta())

# pass in a single taxon
res <- sp_bhl_ocr(x$`allium falcifolium`)
sp_bhl_ocr(x$`castilleja rubicundula`)

# or many taxa
sp_bhl_ocr(x[1:2])

# or all of them at once
sp_bhl_ocr(x)
}
}
