% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_bhl_save.R
\name{sp_bhl_save}
\alias{sp_bhl_save}
\title{Save OCR corpus}
\usage{
sp_bhl_save(x, dir_path = ".", taxon = NULL)
}
\arguments{
\item{x}{An object of class \code{bhl_ocr} from a call to \code{\link{sp_bhl_ocr}},
or a list of such objects. if a list, can be named or unnamed.}

\item{dir_path}{base directory to put files into. Default: \code{.}, current working directory}
}
\description{
Save OCR corpus
}
\details{
Each object passed in has their OCR'ed text blobs written into a folder, where
each file is a separate OCR blob, named by the page ID for that OCRed blob. If pass in a
named object with the taxon name, the folder uses the taxon name - if no names on objects
then we give each folder a random folder name.
}
\examples{
\dontrun{
geom <- 'POLYGON((-124.07 41.48,-119.99 41.48,-119.99 35.57,-124.07 35.57,-124.07 41.48))'
res <- sp_occ_gbif(geometry = geom, limit = 5)
(x <- res \%>\% sp_list() \%>\% sp_bhl_meta() \%>\% .[1:3] \%>\% sp_bhl_ocr())

# write all to disk, each element of list to a separate directory
## with names
sp_bhl_save(x)
## without names
sp_bhl_save(unname(x))

# you can just save some ocr'ed elements
x[1:2] \%>\% sp_bhl_save()

# just one
## single index includes name
sp_bhl_save(x[1])
## double index loses name
sp_bhl_save(x[[1]])
}
}

