% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_lit_names.R
\name{sp_lit_names}
\alias{sp_lit_names}
\title{Extract scientific names from articles}
\usage{
sp_lit_names(x, progress = TRUE, ...)
}
\arguments{
\item{x}{An object of class \code{sp_lit_text} or \code{sp_lit_text_one}}

\item{progress}{(logical) print a progress bar. default: \code{TRUE}}

\item{...}{arguments passed on to \code{namext::name_extract()}}
}
\value{
a data.frame
}
\description{
Depends on non-CRAN package namext
}
\details{
make sure to install gnfinder first
\url{https://github.com/gnames/gnfinder}
}
\examples{
\dontrun{
library(spocc)
taxa <- c('Pinus contorta', 'Accipiter striatus')
res <- occ(query=taxa, from = c("gbif", "bison"), limit = 15)
w <- sp_lit_meta(x = res, from = c("pubmed", "bhl"))
out <- sp_lit_text(x = w[[1]], from = c("pubmed", "bhl"))
sp_lit_names(x = out$bhl$OcrText)
sp_lit_names(x = out$bhl)
}
}
