% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_lit_text.R
\name{sp_lit_text}
\alias{sp_lit_text}
\title{Fetch full text of articles}
\usage{
sp_lit_text(x, from = "pubmed", progress = TRUE)
}
\arguments{
\item{x}{An object of class \code{lit_meta} or \code{lit_meta_one}}

\item{from}{(character) one or more of: bhl, pubmed (default: pubmed)}

\item{progress}{(logical) print a progress bar. default: \code{TRUE}}
}
\value{
An object of class \code{lit_text}, or a list of such objects
}
\description{
Fetch full text of articles
}
\details{
See \link{spplit_auth} for authentication
}
\examples{
\dontrun{
geom <- 'POLYGON((-124.07 41.48,-119.99 41.48,-119.99 35.57,-124.07 35.57,-124.07 41.48))'
res <- sp_occ_gbif(geometry = geom)
(x <- sp_lit_meta(sp_spp(res)[1:2]))

# pass in a one taxon
res <- sp_lit_text(x$`allium falcifolium`)
sp_lit_text(x$`castilleja rubicundula`)

# or many taxa
sp_lit_text(x[1:2])

# or all of them at once
sp_lit_text(x)
}
}
