% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_occ_gbif.R
\name{sp_occ_gbif}
\alias{sp_occ_gbif}
\title{Search for occurrences in GBIF}
\usage{
sp_occ_gbif(query = NULL, geometry = NULL, limit = 10,
  cas_coll = "botany", args = NULL, ...)
}
\arguments{
\item{query}{(character) Scientific names, optional}

\item{geometry}{WKT or bounding box, optional}

\item{limit}{(integer) Number results to return}

\item{cas_coll}{(character) CAS collection name OR a collection code. See Details.}

\item{args}{additional args to GBIF, see \code{\link[rgbif]{occ_search}} for
the options. Any parameters you can pass to \code{occ_search} you can pass to
this parameter.}

\item{...}{Further options passed to \code{\link[spocc]{occ}}}
}
\description{
Search for occurrences in GBIF
}
\details{
This has hard-coded internal settings to get data from the
Cal Academy of Sciences collections within GBIF
}
\section{cas_coll options}{

\itemize{
 \item entomology
 \item herpetology
 \item ichthyology
 \item invertebrate zoology & geology
 \item ornithology
 \item mammalogy
 \item botany
}

OR a dataset key, see \code{\link{gbif_datasets}}
}
\examples{
\dontrun{
geom <- 'POLYGON((-124.07 41.48,-119.99 41.48,-119.99 35.57,-124.07 35.57,-124.07 41.48))'
geom <- 'POLYGON((30.1 10.1, 10 20, 20 60, 60 60, 30.1 10.1))'

res <- sp_occ_gbif(geometry = geom)
res \%>\% sp_list()
x <- res \%>\% sp_list() \%>\% sp_bhl_meta()
ocred <- x[1:10] \%>\% sp_bhl_ocr
ocred \%>\% sp_bhl_save()
#res \%>\% sp_list() \%>\% sp_bhl_meta() \%>\% sp_bhl_ocr \%>\% sp_bhl_save()

sp_occ_gbif(geometry = geom, args = c(basisOfRecord = 'PRESERVED_SPECIMEN'))
sp_occ_gbif(geometry = geom, args = c(basisOfRecord = 'OBSERVATION'))
sp_occ_gbif(geometry = geom, args = c(basisOfRecord = 'HUMAN_OBSERVATION'))

# specify CAS collection (default: botany)
sp_occ_gbif(geometry = geom, cas_coll = "entomology")
## many
res <- sp_occ_gbif(cas_coll = c("entomology", "botany"))
res$gbif$data
## all
sp_occ_gbif(cas_coll = "all")
}
}

