% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spplit-package.R
\docType{package}
\name{spplit-package}
\alias{spplit}
\alias{spplit-package}
\title{Find related literature data for species occurrences.}
\description{
Find related literature data for species occurrences.
}
\keyword{package}

